refactor: 코파일럿 리뷰 제안에 따른 코드 개선
