+++ src/features/user/hooks/useUserQueries.ts
@@ -226,56 +226,3 @@ export function useUpdateProfile(
     ...options,
   });
 }
-
-/**
- * 인증 상태를 종합적으로 관리하는 컴포지션 훅
- */
-export function useAuth() {
-  const queryClient = useQueryClient();
-
-  const profileQuery = useUserProfile();
-  const roleQuery = useUserRole();
-
-  const isLoggedIn = roleQuery.data !== null && roleQuery.data !== undefined;
-  const needsProfileCompletion = roleQuery.data === "PRE_REGISTER";
-  const userProfile = needsProfileCompletion ? null : profileQuery.data;
-
-  const isLoading =
-    roleQuery.isLoading || (isLoggedIn && !needsProfileCompletion && profileQuery.isLoading);
-
-  const login = (userData: UserProfileResponse, userRole: string) => {
-    queryClient.setQueryData(USER_QUERY_KEYS.profile(), userData);
-    queryClient.setQueryData(USER_QUERY_KEYS.role(), userRole);
-  };
-
-  const logout = async () => {
-    try {
-      await userService.logout();
-    } catch (logoutError) {
-      console.error("백엔드 로그아웃 API 호출 실패:", logoutError);
-    } finally {
-      queryClient.removeQueries({ queryKey: USER_QUERY_KEYS.profile() });
-      queryClient.removeQueries({ queryKey: USER_QUERY_KEYS.role() });
-      deleteCookie("authorization", { path: "/", domain: ".gamzatech.site" });
-    }
-  };
-
-  const refetchAuthStatus = async () => {
-    const results = await Promise.allSettled([roleQuery.refetch(), profileQuery.refetch()]);
-    return results;
-  };
-
-  return {
-    isLoggedIn,
-    userProfile,
-    needsProfileCompletion,
-    isLoading,
-    error: roleQuery.error || profileQuery.error,
-    isError: roleQuery.isError || profileQuery.isError,
-    login,
-    logout,
-    refetchAuthStatus,
-    profileQuery,
-    roleQuery,
-  };
-}
